Add keyboard shortcuts for rebasing
When doing a rebase, git lists a number of verbs (pick, drop, edit, ...)
that can be applied to each change in the rebase.

All these verbs come with a single letter abbreviation.

With this change in place, in Git Rebase Message mode, pressing that
single letter on a line will replace the verb on that line with the one
you indicated.

These shortcuts are active only when:
* You are editing a file in Git Rebase Message mode
* You are on a line denoting a change

For other lines in the rebase messages, like comments or exec lines, all
keys work as usual.

I have tested this by having a buffer in Git Rebase Message mode and
typing on commit lines, exec lines, empty lines and comment lines. I
have successfully tested typing both lowercase and uppercase letters on
the non-commit lines.